update VSP header file to VIED 2.19 release
BZ: 190671

update VSP header file to VIED 2.19 release

Category: feature differentiation
Domain: Video.Media-Video Codecs
Origin: internal
Upstream-Candidate: no, proprietary

+++ original/linux/vsp_fw.h
@@ -574,7 +574,7 @@ struct VssVp8encSequenceParameterBuffer {
 	uint32_t ts_target_bitrate[3];
 	uint32_t ts_rate_decimator[3];
 	uint32_t ts_periodicity;
-	uint32_t ts_layer_id[32];
+	uint8_t ts_layer_id[32];
 	struct VssProcPictureVP8 ref_frame_buffers[4];
 };
